% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_vcf.R
\name{bam_to_vcf}
\alias{bam_to_vcf}
\title{Convert `.bam`-Like Data to `.vcf` Format}
\usage{
bam_to_vcf(
  patient_id,
  sample_id,
  output_dir = tempdir(),
  vcf_file = paste0(output_dir, "/", patient_id, "_", sample_id, ".vcf")
)
}
\arguments{
\item{patient_id}{A character string specifying the patient ID. This will be used as the base name for the output `.vcf` file.}

\item{sample_id}{A character string specifying the sample ID. This will be used as the base name for the output `.vcf` file.}

\item{output_dir}{A character string specifying the directory where the `.vcf` file will be saved.}

\item{vcf_file}{A character string specifying the path to the output `.vcf` file. Default is "dummy.vcf".}
}
\description{
This function simulates the conversion of `.bam`-like data to `.vcf` format.
It creates a dummy `.vcf` file with dummy data.
}
\details{
The function creates a dummy `.vcf` file with dummy data.
The `.vcf` file contains the following fields:
- CHROM: Chromosome name
- POS: Position on the chromosome
- ID: Variant identifier (dummy value ".")
- REF: Reference allele
- ALT: Alternate allele
- QUAL: Quality score
- dg: Total depth
- gt: Genotype
The function writes the `.vcf` file to the specified path.
}
\examples{
# Convert `.bam`-like data to `.vcf` format
patient_id <- "patient_123"
sample_id <- "smp_001"
output_dir <- tempdir()
bam_to_vcf(patient_id, sample_id, output_dir, vcf_file = paste0(output_dir,
"/", patient_id, ".vcf"))
}
