% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{clean}
\alias{clean}
\title{Function to clean the data file}
\usage{
clean(x, y)
}
\arguments{
\item{x}{An integer.}

\item{y}{An integer.}
}
\value{
A modified data frame.
}
\description{
Function to clean the data file
}
\examples{
x <- 3
y <- 5
clean(x, y)
}
