% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fastq_file.R
\name{create_fastq_file}
\alias{create_fastq_file}
\title{Create an empty .fastq file using patient ID}
\usage{
create_fastq_file(patient_id)
}
\arguments{
\item{patient_id}{A vector of patient IDs}
}
\value{
An empty .fastq file with the patient ID in the name
}
\description{
Create an empty .fastq file using patient ID
}
\examples{
patient_id <- "patient_1"
create_fastq_file(patient_id)
}
