% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file.R
\name{create_file}
\alias{create_file}
\title{Title}
\usage{
create_file(n_pat, n_samp)
}
\arguments{
\item{n_pat}{Number of patients}

\item{n_samp}{Number of samples}
}
\value{
A vector of file names
}
\description{
Title
}
\examples{
n_pat <- 3
n_samp <- 3
create_file(n_pat,n_samp)
}
