% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_function.R
\name{data_frame}
\alias{data_frame}
\title{Title}
\usage{
data_frame(x, y, z)
}
\arguments{
\item{x}{Column 1 of data frame}

\item{y}{Column 2 of data frame}

\item{z}{Column 3 of data frame}
}
\value{
A data frame
}
\description{
Title
}
\examples{
nrow <- 2
ncol <- 3
x <- c("a","p")
y <- c(12.56,38)
z <- c("abc","pqr")
data_frame(x, y, z)
}
