% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_fastq.R
\name{fill_fastq}
\alias{fill_fastq}
\title{Function to Fill a `.fastq` File with Random Reads}
\usage{
fill_fastq(patient_id, sample_id, output_dir = tempdir(), n, read_length)
}
\arguments{
\item{patient_id}{A character string representing the patient ID.}

\item{sample_id}{A character string representing the sample ID.}

\item{output_dir}{A character string specifying the directory where the `.fastq` file will be saved.}

\item{n}{An integer specifying the number of reads to generate.}

\item{read_length}{An integer specifying the length of each read. It should be multiple of 4. Default is 8.}
}
\description{
This function populates a `.fastq` file with random reads for a given patient ID.
The function first creates an empty `.fastq` file using the patient ID,
then generates a specified number of random reads,
and writes them to the file.
}
\details{
The function first checks if the `.fastq` file already exists.
If it does, it appends the new reads to the existing file.
If not, it creates a new file.
The generated reads are in `.fastq` format, which consists of four lines per read:
\itemize{
\item Line 1: Sequence identifier (e.g., @patient_id_read1)
\item Line 2: Random DNA sequence
\item Line 3: Separator (e.g., +)
\item Line 4: Quality scores
}
}
\examples{
# Fill a `.fastq` file for patient ID "patient_123", sample ID "smp_001" with 2 reads of length 8
fill_fastq("patient_001", "smp_001", output_dir = tempdir(), n = 2, read_length = 8)

}
