% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreads.R
\name{rreads}
\alias{rreads}
\title{Function to Generate a Random Sample of Reads in `.fastq` Format}
\usage{
rreads(patient_id, sample_id, n, read_length = 8)
}
\arguments{
\item{patient_id}{A character string representing the patient ID.}

\item{sample_id}{A character string representing the sample ID.}

\item{n}{An integer specifying the number of reads to generate.}

\item{read_length}{An integer specifying the length of each read. It should be multiple of 4. Default is 8.}
}
\value{
A character vector containing the generated `.fastq` lines.
}
\description{
This function generates a specified number of random reads in `.fastq` format.
Each read consists of a sequence identifier, a random DNA sequence, a separator line, and quality scores.
}
\details{
The function generates random DNA sequences using the nucleotides A, T, G, and C.
The quality scores are represented as ASCII characters, with 'I' indicating high quality.
The output is formatted as a `.fastq` file, which consists of four lines per read:
\itemize{
 \item Line 1: Sequence identifier (e.g., @patient_id_read1)
 \item Line 2: Random DNA sequence
 \item Line 3: Separator (e.g., +)
 \item Line 4: Quality scores
 }
 @seealso
 \code{\link{create_empty_fastq}} for creating an empty `.fastq` file.
}
\examples{
# Generate 5 random reads for patient ID "patient_001" and sample ID
# "smp_001"
rreads("patient_001", "smp_001", n = 5)
# Generate 10 random reads for patient ID "patient_001" and sample ID
# "smp_002" with a read length of 12
rreads("patient_001", "smp_002", n = 10, read_length = 12)
}
