% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_vcf.R
\name{bam_to_vcf}
\alias{bam_to_vcf}
\title{Convert BAM-like data to VCF format}
\usage{
bam_to_vcf(patient_id, vcf_file = "dummy.vcf")
}
\arguments{
\item{patient_id}{A character string specifying the patient ID. This will be used as the base name for the output VCF file.}

\item{vcf_file}{A character string specifying the path to the output VCF file. Default is "dummy.vcf".}
}
\description{
This function simulates the conversion of BAM-like data to VCF format.
It creates a mock VCF file with dummy data.
}
\details{
The function creates a mock VCF file with dummy data.
The VCF file contains the following fields:
- CHROM: Chromosome name
- POS: Position on the chromosome
- ID: Variant identifier (dummy value ".")
- REF: Reference allele
- ALT: Alternate allele
- QUAL: Quality score
- dg: Total depth
- gt: Genotype
The function writes the VCF file to the specified path.
}
\examples{
# Convert BAM-like data to VCF format
# bam_to_vcf("output.vcf")
patient_id <- "patient_123"
bam_to_vcf(patient_id = patient_id, vcf_file ="output.vcf")

}
