% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_fastq.R
\name{create_empty_fastq}
\alias{create_empty_fastq}
\title{Create an empty .fastq file using patient ID}
\usage{
create_empty_fastq(patient_id, sample_id, output_dir = tempdir())
}
\arguments{
\item{patient_id}{A character vector representing the patient IDs.}

\item{sample_id}{A character vector representing the sample IDs.}

\item{output_dir}{A character string specifying the directory where the `.fastq` file will be saved.}
}
\description{
This function creates an empty `.fastq` file with the specified patient ID and sample ID.
The file will contain a single entry with the patient ID as the sequence identifier.
The sequence and quality scores will be empty
}
\details{
The function constructs a `.fastq` file with the following format:
\itemize{
 \item Line 1: Sequence ID (e.g., @patient_id)
 \item Line 2: Empty sequence
 \item Line 3: Separator (e.g., +)
 \item Line 4: Empty quality scores
 }
}
\examples{
# Create an empty `.fastq` file for patient ID "Patient_001" and
#  sample ID "SMP_001"
 create_empty_fastq("Patient_001", "SMP_001")
# Inside a specific directory, Create an empty `.fastq` file for
# patient ID "Patient_001" and sample ID "SMP_002"
 create_empty_fastq("Patient_001", "SMP_002", output_dir = tempdir())
}
