% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create metadata for cBioPortal study}
\usage{
create_metadata(patient_id)
}
\arguments{
\item{patient_id}{A vector of patient IDs to be included in the clinical data.}
}
\description{
This function creates a directory structure and metadata files for a cBioPortal study.
It generates clinical patient data, mutation data, and associated metadata files.
The function creates a directory named "cbioportal_study" and writes the following files:
}
\details{
- data_clinical_patient.txt: Contains clinical patient data.
- meta_clinical_patient.txt: Metadata for clinical patient data.
- meta_study.txt: Metadata for the study.
- data_mutations.txt: Contains mutation data.
- meta_mutations.txt: Metadata for mutation data.
}
\examples{
# create_metadata(c("P1", "P2"))
}
