% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_fastq.R
\name{fill_fastq}
\alias{fill_fastq}
\title{Function to fill a FASTQ file with random reads}
\usage{
fill_fastq(patient_id, output_dir, n, read_length)
}
\arguments{
\item{patient_id}{A character string representing the patient ID.}

\item{output_dir}{A character string specifying the directory where the FASTQ file will be saved. Default is the current working directory.}

\item{n}{An integer specifying the number of reads to generate.}

\item{read_length}{An integer specifying the length of each read. Default is 8. It should be multiple of 4.}
}
\description{
This function populates a FASTQ file with random reads for a given patient ID.
The function first creates an empty FASTQ file using the patient ID,
then generates a specified number of random reads,
and writes them to the file.
}
\details{
The function first checks if the FASTQ file already exists.
If it does, it appends the new reads to the existing file.
If not, it creates a new file.
The generated reads are in FASTQ format, which consists of four lines per read:
\itemize{
\item Line 1: Sequence identifier (e.g., @patient_id_read1)
\item Line 2: Random DNA sequence
\item Line 3: Separator (e.g., +)
\item Line 4: Quality scores
}
}
\examples{
# Fill a FASTQ file for patient ID "patient_123" with 2 reads of length 8
fill_fastq("patient_123",output_dir = ".", n = 2, read_length = 8)

}
